Bug fix: Add easytags_file to &tags option
While writing the `easytags.vim` plug-in at one point I added the
following line to my `~/.vimrc` script:

    :set tags=~/.vimtags

After publishing the plug-in I totally forgot about the `tags` option
and how the plug-in depends on it for all of its functionality :-S

The plug-in now adds the (possibly user-defined) value of the
`easytags_file` option to Vim's `tags` option, thereby making Vim's
internals aware of the tags file.

Oh by the way, the plug-in is now more-or-less Windows compatible.
The only remaining problems are the flashing command prompt (which
I know how to fix) and the fact that CursorHold seems to fire
repeatedly...