Automatic/optional integration with shell.vim
I developed (parts of) my shell.vim plug-in after noticing the annoying
command prompt windows that pop up temporarily on Windows when the
easytags plug-in runs ctags. Because the shell.vim plug-in was a proof
of concept I hadn't published any integration between the two plug-ins.
Since then I've had several requests for this so here it is :-)