FIX: Avoid error about empty Dictionary key.
I had a degenerate tags file that had *two* tabs separating the tag from the filespec column; parsing that yields an empty filespec, which caused E713. There used to be an explicit check for that; with your recent refactorings, one instance of that check was lost. Here it is again!